requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0  
anyio==3.6.2      
appdirs==1.4.4    
APScheduler==3.6.3
argcomplete==2.0.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astroid==2.12.12
asttokens==2.0.5
async-generator==1.10
async-timeout==4.0.2
atomicwrites==1.4.1
attrs==22.1.0
Automat==22.10.0
Babel==2.11.0
backcall==0.2.0
base58==2.1.1
beautifulsoup4==4.11.1
bitarray==2.6.0
black==22.10.0
bleach==5.0.1
Brotli==1.0.9
bs4==0.0.1
cachetools==4.2.2
certifi==2022.9.24
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
coloredlogs==14.3
constantly==15.1.0
contourpy==1.0.6
cryptography==37.0.2
cssselect==1.2.0
cycler==0.11.0
Cython==0.29.30
cytoolz==0.12.0
dataclassy==0.11.1
debugpy==1.6.3
decorator==4.4.2
defusedxml==0.7.1
dill==0.3.6
discord.py==1.7.3
distlib==0.3.6
docopt==0.6.2
DT-Scraper==1.0.4
eip712==0.1.0
entrypoints==0.4
eth-abi==2.2.0
eth-account==0.5.9
eth-brownie==1.19.2
eth-event==1.2.3
eth-hash==0.3.3
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.10.0
execnet==1.9.0
executing==1.0.0
ez-setup==0.9
fake-useragent==0.1.14
Faker==13.11.1
fastapi==0.86.0
fastjsonschema==2.16.2
filelock==3.7.0
Flask==2.1.2
fonttools==4.38.0
frozenlist==1.3.1
fs==2.4.16
google-api-core==2.10.1
google-api-python-client==2.61.0
google-auth==2.11.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.3
googleapis-common-protos==1.56.4
greenlet==1.1.3
h11==0.12.0
hexbytes==0.2.3
httpcore==0.15.0
httplib2==0.20.4
httpx==0.23.0
humanfriendly==10.0
hyperlink==21.0.0
hypothesis==6.27.3
idna==3.4
imageio==2.22.1
imageio-ffmpeg==0.4.7
importlib-metadata==5.0.0
incremental==22.10.0
inflection==0.5.0
iniconfig==1.1.1
insta-scrape==2.1.2
instaloader==4.9.3
instalooter==2.4.4
ipfshttpclient==0.8.0a2
ipykernel==6.15.2
ipython==8.5.0
ipython-genutils==0.2.0
isort==5.10.1
itemadapter==0.7.0
itemloaders==1.0.6
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.1
json5==0.9.10
jsonschema==3.2.0
jupyter-core==4.11.1
jupyter-server==1.23.2
jupyter_client==7.3.5
jupyterlab==3.5.0
jupyterlab-pygments==0.2.2
jupyterlab_server==2.16.3
kiwisolver==1.4.4
lazy-object-proxy==1.7.1
limits==1.6
lru-dict==1.1.8
lxml==4.9.1
Mako==1.2.3
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mccabe==0.7.0
mistune==2.0.4
moviepy==1.0.3
mpl-finance==0.10.1
multiaddr==0.0.9
multidict==6.0.2
multitasking==0.0.11
mutagen==1.45.1
mypy-extensions==0.4.3
mythx-models==1.9.1
nbclassic==0.4.8
nbclient==0.7.0
nbconvert==7.2.5
nbformat==5.7.0
nest-asyncio==1.5.5
netaddr==0.8.0
networkx==2.8.7
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.22.3
oauth2client==4.1.3
oauthlib==3.2.1
opencv-python==4.6.0.66
orjson==3.8.1
outcome==1.1.0
packaging==21.3
pandas==1.4.2
pandas-datareader==0.10.0
pandocfilters==1.5.0
parse==1.19.0
parsel==1.7.0
parsimonious==0.8.1
parso==0.8.3
pathspec==0.10.1
pbr==5.11.0
pdfkit==1.0.0
pdoc3==0.10.0
pickleshare==0.7.5
Pillow==9.2.0
pipenv==2022.5.2
pipx==1.1.0
platformdirs==2.5.2
playwright==1.26.1
pluggy==1.0.0
proglog==0.1.10
prometheus-client==0.15.0
prompt-toolkit==3.0.31
Protego==0.2.1
protobuf==3.19.5
psutil==5.9.2
pure-eval==0.2.2
py==1.11.0
py-solc-ast==1.2.9
py-solc-x==1.1.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.15.0
pycryptodomex==3.15.0
pydantic==1.10.2
PyDispatcher==2.0.6
pyee==8.1.0
pyflakes==2.5.0
Pygments==2.13.0
pygments-lexer-solidity==0.7.0
PyJWT==1.7.1
pylint==2.15.5
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyppeteer==1.0.2
pyquery==1.4.3
pyreadline3==3.4.1
pyrsistent==0.18.1
PySocks==1.7.1
pytest==6.2.5
pytest-forked==1.4.0
pytest-xdist==1.34.0
python-dateutil==2.8.2
python-dotenv==0.16.0
python-telegram-bot==13.14
pythx==1.6.1
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pyvim==3.0.3
PyWavelets==1.4.1
pywebio==1.7.0
pywin32==304
pywinpty==2.0.9
PyYAML==5.4.1
pyzmq==23.2.1
queuelib==1.6.2
requests==2.28.1
requests-file==1.5.1
requests-html==0.10.0
requests-oauthlib==1.3.1
retrying==1.3.3
rfc3986==1.5.0
rlp==2.0.1
rsa==4.9
scikit-image==0.19.3
scipy==1.9.2
Scrapy==2.7.1
selenium==4.4.0
selenium-stealth==1.0.6
semantic-version==2.10.0
Send2Trash==1.8.0
service-identity==21.1.0
six==1.16.0
slowapi==0.1.6
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
stack-data==0.5.0
starlette==0.20.4
stevedore==4.1.0
stockstats==0.4.1
telegram-send==0.34
tenacity==8.1.0
terminado==0.17.0
tifffile==2022.10.10
TikTokApi==5.2.2
tinycss2==1.2.1
tldextract==3.4.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.6
toolz==0.12.0
tornado==6.1
tqdm==4.64.1
traitlets==5.3.0
trio==0.20.0
trio-websocket==0.9.2
Twisted==22.10.0
twisted-iocpsupport==1.0.2
typing_extensions==4.4.0
tzdata==2022.6
tzlocal==4.2
ua-parser==0.16.1
uritemplate==4.1.1
urllib3==1.26.12
user-agents==2.2.0
userpath==1.8.0
uvicorn==0.19.0
validate-email==1.3
varint==1.0.2
verboselogs==1.7
virtualenv==20.16.5
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
vvm==0.1.0
vyper==0.3.7
w3lib==2.0.1
wcwidth==0.2.5
web3==5.31.1
webdriver-manager==3.7.0
webencodings==0.5.1
websocket-client==1.4.2
websockets==10.4
Werkzeug==2.1.2
wrapt==1.14.1
wsproto==1.1.0
yarl==1.8.1
yfinance==0.1.86
yt-dlp==2022.10.4
zipp==3.10.0
zope.interface==5.5.1